+++ chaincode/resource_types/resource_types.go
@@ -26,24 +26,3 @@ type ResourceTypesContract struct {
 func (rc *ResourceTypesContract) InitLedger(ctx contractapi.TransactionContextInterface) error {
 	return nil
 }
-
-func (rc *ResourceTypesContract) SetPrivateData(
-	ctx contractapi.TransactionContextInterface,
-	id string,
-	privateName string,
-) error {
-	mspID, err := ctx.GetClientIdentity().GetMSPID()
-	if err != nil {
-		return err
-	}
-
-	if err = ctx.GetStub().PutPrivateData(
-		mspID+"ResourceTypesPrivateData",
-		id,
-		[]byte(privateName),
-	); err != nil {
-		return err
-	}
-
-	return nil
-}
